Fan profiles in OpenBMC supports Acoustic and performance profiles only If no profiles are fully supported across all configured fan domains, the BMC defaults to performance profile
If we add a "Profiles" vector into the PID and Stepwise entity-manager configurations the logic becomes much more simple. We default to all profiles, and can limit the scope of some configurations to the selected profile.
On start dbus configuration will look for profiles.  If any exist it will attempt to find the thermal mode interface: This can be anywhere on d-bus, tested placement was in settingsd. Based on the selected profile(s) it will create a 'whitelist' of controllers, and will remove any not in that whitelist.

Tested :

Added and removed zone configuration from thermal ,based on acoustic or
performance mode

Signed-off-by: thrishalig <thrishalig@ami.com>