Added proper path for retrieving cupsensors
Description:
The sensor path in the bhs platform was changed, causing the sensors to not be listed correctly in redfish.

Before fix:
Cup sensors were not appearing in the redfish cups sensor list.

After fix:
Cup sensors are now properly listed in redfish.

Tested:
Successfully build and secured images generated.